requirements.txt
python-dotenv
openai
pyupbit
pandas
pandas_ta
schedule
python-telegram-bot==13.14
asyncio